+++ statistics-sdk/src/main/java/com/dwarfeng/statistics/sdk/bean/entity/JSFixedFastJsonHistoryTask.java
@@ -17,7 +17,7 @@ import java.util.Objects;
  */
 public class JSFixedFastJsonHistoryTask implements Bean {
 
-    private static final long serialVersionUID = -6129487410363904573L;
+    private static final long serialVersionUID = -4305845389083251049L;
 
     public static JSFixedFastJsonHistoryTask of(HistoryTask historyTask) {
         if (Objects.isNull(historyTask)) {
@@ -33,7 +33,6 @@ public class JSFixedFastJsonHistoryTask implements Bean {
                     historyTask.getDuration(),
                     historyTask.getExpiredDate(),
                     historyTask.getDiedDate(),
-                    historyTask.getExecutionNodeId(),
                     historyTask.getMessage()
             );
         }
@@ -66,10 +65,7 @@ public class JSFixedFastJsonHistoryTask implements Bean {
     @JSONField(name = "died_date", ordinal = 9)
     private Date diedDate;
 
-    @JSONField(name = "execution_node_id", ordinal = 10)
-    private int executionNodeId;
-
-    @JSONField(name = "message", ordinal = 11)
+    @JSONField(name = "message", ordinal = 10)
     private String message;
 
     public JSFixedFastJsonHistoryTask() {
@@ -77,7 +73,7 @@ public class JSFixedFastJsonHistoryTask implements Bean {
 
     public JSFixedFastJsonHistoryTask(
             JSFixedFastJsonLongIdKey key, JSFixedFastJsonLongIdKey statisticsSettingKey, int status, Date createDate,
-            Date startDate, Date endDate, Long duration, Date expiredDate, Date diedDate, int executionNodeId,
+            Date startDate, Date endDate, Long duration, Date expiredDate, Date diedDate,
             String message
     ) {
         this.key = key;
@@ -89,7 +85,6 @@ public class JSFixedFastJsonHistoryTask implements Bean {
         this.duration = duration;
         this.expiredDate = expiredDate;
         this.diedDate = diedDate;
-        this.executionNodeId = executionNodeId;
         this.message = message;
     }
 
@@ -165,14 +160,6 @@ public class JSFixedFastJsonHistoryTask implements Bean {
         this.diedDate = diedDate;
     }
 
-    public int getExecutionNodeId() {
-        return executionNodeId;
-    }
-
-    public void setExecutionNodeId(int executionNodeId) {
-        this.executionNodeId = executionNodeId;
-    }
-
     public String getMessage() {
         return message;
     }
@@ -193,7 +180,6 @@ public class JSFixedFastJsonHistoryTask implements Bean {
                 ", duration=" + duration +
                 ", expiredDate=" + expiredDate +
                 ", diedDate=" + diedDate +
-                ", executionNodeId=" + executionNodeId +
                 ", message='" + message + '\'' +
                 '}';
     }